+++ pandapower/plotting/__init__.py
@@ -8,17 +8,19 @@ from pandapower.plotting.geo import *
 from pandapower.plotting.to_html import to_html
 
 import types
-from matplotlib.backend_bases import GraphicsContextBase, RendererBase
 
+try:
+    from matplotlib.backend_bases import GraphicsContextBase, RendererBase
 
-class GC(GraphicsContextBase):
-    def __init__(self):
-        super().__init__()
-        self._capstyle = 'round'
+    class GC(GraphicsContextBase):
+        def __init__(self):
+            super().__init__()
+            self._capstyle = 'round'
 
+    def custom_new_gc(self):
+        return GC()
 
-def custom_new_gc(self):
-    return GC()
+    RendererBase.new_gc = types.MethodType(custom_new_gc, RendererBase)
 
-
-RendererBase.new_gc = types.MethodType(custom_new_gc, RendererBase)
+except ImportError:
+    pass